cli: adopt static-help split for avatar, backup, bash, and browser
Third batch of the static-help migration (after attachment, audit, auth).
These four needed contract growth in `CliCommandHelp`:

- Nested subcommand groups: `CliSubcommandHelp` gains a recursive
  `subcommands` field, applied recursively by `applyCommandHelp`
  (avatar `character …`, backup `destinations …`, browser `tabs …`).
- Positional arguments: `args` on commands and subcommands keeps `name`
  bare (dedup/slugs stay parseable) while the new `commandSpec()` helper
  builds the Commander registration string (`bash <command>`,
  `destinations add <path>`, `set-encrypt <path> <value>`).
- Enum options: `choices` on `CliOptionHelp`, applied via Commander's
  `Option.choices()` (browser `--browser-mode`).

The memory renderer (`cli-command-content.ts`) walks the subcommand tree
depth-first, heading each section with the full command path a user would
type (`backup destinations add <path> — …`), and annotates option lines
with (required)/(choices: …)/(default: …).

`browser.help.ts` is derived from `BROWSER_OPERATION_META` rather than
hand-written: the operations contract stays the single source of truth
for the ~17 per-operation subcommands, so the declarative help cannot
drift from what the daemon supports. It carries the same sanctioned
`cli/no-daemon-internals` disable as `browser.ts` (registration-time
data leaf; noted in AGENTS.md). `tabs select/close` previously coerced
`--tab-id` with a Commander parser fn; the declarative contract is pure
data, so the `Number()` coercion moves into the actions (same NaN
semantics on invalid input).

Verified byte-for-byte identical `--help` output across all 29 command/
subcommand screens against the pre-split baseline.
