+++ assistant/src/cli/index.help.ts
@@ -14,10 +14,18 @@
 import { attachmentHelp } from "./commands/attachment.help.js";
 import { auditHelp } from "./commands/audit.help.js";
 import { authHelp } from "./commands/auth.help.js";
+import { avatarHelp } from "./commands/avatar.help.js";
+import { backupHelp } from "./commands/backup.help.js";
+import { bashHelp } from "./commands/bash.help.js";
+import { browserHelp } from "./commands/browser.help.js";
 import type { CliCommandHelp } from "./lib/cli-command-help.js";
 
 export const CLI_COMMAND_HELP: readonly CliCommandHelp[] = [
   attachmentHelp,
   auditHelp,
   authHelp,
+  avatarHelp,
+  backupHelp,
+  bashHelp,
+  browserHelp,
 ];
